Add Server module
This is the module that will serve NMEA data from the GPS to TCP clients.
Currently the TCP host and port are hardcoded. Next step would be to run
it on dynamic port on all IP addresses and then use Avahi to announce the
service to Geoclue can use it.